+++ src/Json/Create/CreateBaseJsonResponse.php
@@ -11,9 +11,9 @@ use Psr\Http\Message\ResponseInterface as Response;
 const FAILURE_RESPONSE_DEFAULT = "{\"success\":false,\"data\":null,\"errors\":[{\"code\":\"server\",\"path\":[],\"message\":\"Internal server error.\"}]}";
 
 /**
- * Base functions to create JSON response.
+ * Create a base JSON response.
  */
-class CreateBaseJsonResponseFunctions extends CreateBaseResponse
+class CreateBaseJsonResponse extends CreateBaseResponse
 {
     protected JsonResponse $json;