+++ src/Json/CreateJsonResponse.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Jder\Psr\Json;
 
-use Jder\Psr\Json\Functions\CreateFailureJsonResponseFunctions;
-use Jder\Psr\Json\Functions\CreateSuccessJsonResponseFunctions;
+use Jder\Psr\Json\Functions\CreateFailureJsonResponse;
+use Jder\Psr\Json\Functions\CreateSuccessJsonResponse;
 use Psr\Http\Message\ResponseInterface as Response;
 
 /**
@@ -16,16 +16,16 @@ class CreateJsonResponse
     /**
      * Create a success JSON response.
      */
-    public static function success(Response $response): CreateSuccessJsonResponseFunctions
+    public static function success(Response $response): CreateSuccessJsonResponse
     {
-        return new CreateSuccessJsonResponseFunctions($response);
+        return new CreateSuccessJsonResponse($response);
     }
 
     /**
      * Create a failure JSON response.
      */
-    public static function failure(Response $response): CreateFailureJsonResponseFunctions
+    public static function failure(Response $response): CreateFailureJsonResponse
     {
-        return new CreateFailureJsonResponseFunctions($response);
+        return new CreateFailureJsonResponse($response);
     }
 }