+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 - rename `ResponseErrorCode` to `ResponseError`
 - rename `CreateResponseBase` to `CreateBaseResponse`
-- rename `CreateJsonBaseResponseFunctions` to `CreateBaseJsonResponseFunctions`
-- rename `CreateJsonSuccessResponseFunctions` to `CreateSuccessJsonResponseFunctions`
-- rename `CreateJsonFailureResponseFunctions` to `CreateFailureJsonResponseFunctions`
+- rename `CreateJsonBaseResponseFunctions` to `CreateBaseJsonResponse`
+- rename `CreateJsonSuccessResponseFunctions` to `CreateSuccessJsonResponse`
+- rename `CreateJsonFailureResponseFunctions` to `CreateFailureJsonResponse`
 
 ## 0.1.0 (2025-12-16)